+++ docs/changelog.rst
@@ -9,7 +9,8 @@ Version 1.0.0
 + 增加 :class:`SuccessReply`
 + 增加 :class:`ImageReply`
 + 删去了 Reply 中过时的 flag
-+ 修复 :class:`werobot.session.FileStorage` 在 PyPy 下的兼容性问题
++ 修复 :class:`werobot.session.filestorage.FileStorage` 在 PyPy 下的兼容性问题
++ 增加 :class:`werobot.session.sqlitestorage.SQLiteStorage`
 
 Version 0.6.1
 ----------------